+++ CMakeLists.txt
@@ -30,6 +30,8 @@ option(WITH_AUDIT "Build with audit support" ON)
 option(WITH_FSVERITY "Build with fsverity support" OFF)
 option(WITH_IMAEVM "Build with IMA support" OFF)
 option(WITH_FAPOLICYD "Build with fapolicyd support" ON)
+option(WITH_SEQUOIA "Build with Sequoia OpenPGP support" ON)
+option(WITH_OPENSSL "Use openssl instead of libgcrypt for internal crypto" OFF)
 option(WITH_INTERNAL_OPENPGP "Use internal OpenPGP parser (DEPRECATED)" OFF)
 option(WITH_READLINE "Build with readline support" ON)
 option(WITH_BZIP2 "Build with bzip2 support" ON)
