+++ rpmio/CMakeLists.txt
@@ -13,13 +13,26 @@ target_include_directories(librpmio PRIVATE
 	${Intl_INCLUDE_DIRS}
 )
 
-if (WITH_INTERNAL_OPENPGP)
-	add_subdirectory(rpmpgp_legacy)
-	target_link_libraries(librpmio PRIVATE rpmpgp_legacy)
-else()
+if (WITH_SEQUOIA)
 	pkg_check_modules(RPMSEQUOIA REQUIRED IMPORTED_TARGET rpm-sequoia>=1.4.0)
 	target_sources(librpmio PRIVATE rpmpgp_sequoia.c)
 	target_link_libraries(librpmio PRIVATE PkgConfig::RPMSEQUOIA)
+else()
+	if (WITH_INTERNAL_OPENPGP)
+		add_subdirectory(rpmpgp_legacy)
+		target_link_libraries(librpmio PRIVATE rpmpgp_legacy)
+	else()
+		target_sources(librpmio PRIVATE rpmpgp_dummy.c)
+	endif()
+	if (WITH_OPENSSL)
+		find_package(OpenSSL 3.0.0 REQUIRED)
+		target_sources(librpmio PRIVATE digest_openssl.c)
+		target_link_libraries(librpmio PRIVATE OpenSSL::Crypto)
+	else()
+		pkg_check_modules(LIBGCRYPT REQUIRED IMPORTED_TARGET libgcrypt)
+		target_sources(librpmio PRIVATE digest_libgcrypt.c)
+		target_link_libraries(librpmio PRIVATE PkgConfig::LIBGCRYPT)
+	endif()
 endif()
 
 set_target_properties(librpmio PROPERTIES
