+++ tests/CMakeLists.txt
@@ -17,10 +17,12 @@ os_release(OS_NAME ID)
 os_release(OS_VERSION VERSION_ID)
 
 set(PYTHON ${Python3_EXECUTABLE})
-if (WITH_INTERNAL_OPENPGP)
+if (WITH_SEQUOIA)
+	set(PGP sequoia)
+elseif (WITH_INTERNAL_OPENPGP)
 	set(PGP internal)
 else()
-	set(PGP sequoia)
+	set(PGP dummy)
 endif()
 set(HAVE_UNSHARE ${HAVE_UNSHARE})
