+++ tests/rpmdb.at
@@ -58,6 +58,7 @@ RPMTEST_CLEANUP
 
 AT_SETUP([rpm -qa and rpmkeys])
 AT_KEYWORDS([rpmdb query])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMDB_INIT
 
 RPMTEST_CHECK([
