+++ tests/rpmpython.at
@@ -216,7 +216,10 @@ myprint(h['arch'])
 [ppc64]
 )
 
-RPMPY_TEST([reading a signed package file 1],[
+AT_SETUP([reading a signed package file])
+AT_KEYWORDS([python])
+AT_SKIP_IF([test x$PGP = xdummy])
+RPMPY_CHECK([
 # avoid rpmlog spew with absolute path to package
 sink = open('/dev/null', 'w')
 rpm.setLogFile(sink)
@@ -231,7 +234,7 @@ sink.close()
 ],
 )
 
-RPMPY_TEST([reading a signed package file 2],[
+RPMPY_CHECK([
 
 with open('${RPMDATA}/keys/rpm.org-rsa-2048-test.pub', 'rb') as keydata_file:
     keydata = keydata_file.read()
@@ -248,6 +251,7 @@ except rpm.error as e:
 ],
 [x86_64]
 )
+RPMTEST_CLEANUP
 
 RPMPY_TEST([add package to transaction],[
 ts.addInstall('${RPMDATA}/RPMS/foo-1.0-1.noarch.rpm', 'u')
