+++ tests/rpmquery.at
@@ -408,6 +408,7 @@ RPMTEST_CLEANUP
 
 AT_SETUP([pgpsig extension query])
 AT_KEYWORDS([query signature])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMTEST_CHECK([
 RPMDB_INIT
 runroot rpm \
