+++ tests/rpmsigdig.at
@@ -172,6 +172,7 @@ RPMTEST_CLEANUP
 # when the key is imported; other implementations should not do this.
 AT_SETUP([rpmkeys --import rsa (rpmdb)])
 AT_KEYWORDS([rpmkeys import])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMTEST_CHECK_UNQUOTED([
 RPMDB_INIT
 
@@ -236,6 +237,7 @@ RPMTEST_CLEANUP
 
 AT_SETUP([rpmkeys --import rsa (fs)])
 AT_KEYWORDS([rpmkeys import])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMTEST_CHECK([
 RPMDB_INIT
 
@@ -287,6 +289,7 @@ RPMTEST_CLEANUP
 # does not support.
 AT_SETUP([rpmkeys --import, signed with a good subkey])
 AT_KEYWORDS([rpmkeys digest signature])
+AT_SKIP_IF([test x$PGP = xdummy])
 AT_SKIP_IF([test x$PGP = xinternal])
 RPMTEST_CHECK([
 RPMDB_INIT
@@ -352,6 +355,7 @@ RPMTEST_CLEANUP
 # Note: the internal PGP implementation does not support expiry.
 AT_SETUP([rpmkeys --import, signed with an expired subkey])
 AT_KEYWORDS([rpmkeys digest signature])
+AT_SKIP_IF([test x$PGP = xdummy])
 AT_SKIP_IF([test x$PGP = xinternal])
 RPMTEST_CHECK([
 RPMDB_INIT
@@ -427,6 +431,7 @@ RPMTEST_CLEANUP
 # Note: the internal PGP implementation does not support revoked subkeys.
 AT_SETUP([rpmkeys --import, signed with a revoked subkey])
 AT_KEYWORDS([rpmkeys digest signature])
+AT_SKIP_IF([test x$PGP = xdummy])
 AT_SKIP_IF([test x$PGP = xinternal])
 RPMTEST_CHECK([
 RPMDB_INIT
@@ -495,6 +500,7 @@ RPMTEST_CLEANUP
 
 AT_SETUP([rpmkeys --import invalid keys])
 AT_KEYWORDS([rpmkeys import])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMDB_INIT
 
 # The following certificates include subkeys with errors.  In general,
@@ -559,6 +565,7 @@ RPMTEST_CLEANUP
 # instead of the key's creation time.
 AT_SETUP([rpmkeys --import different-creation-times])
 AT_KEYWORDS([rpmkeys import])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMDB_INIT
 RPMTEST_CHECK([
 runroot rpmkeys --import /data/keys/different-creation-times.asc
@@ -630,6 +637,7 @@ RPMTEST_CLEANUP
 # Test pre-built package verification
 AT_SETUP([rpmkeys -K <signed> 1])
 AT_KEYWORDS([rpmkeys digest signature])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMTEST_CHECK([
 RPMDB_INIT
 
@@ -648,6 +656,7 @@ RPMTEST_CLEANUP
 # Test pre-built package verification
 AT_SETUP([rpmkeys -Kv <signed> 1])
 AT_KEYWORDS([rpmkeys digest signature])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMTEST_CHECK([
 RPMDB_INIT
 
@@ -723,6 +732,7 @@ RPMTEST_CLEANUP
 # Test pre-built corrupted package verification (corrupted signature)
 AT_SETUP([rpmkeys -Kv <corrupted signed> 1])
 AT_KEYWORDS([rpmkeys digest signature])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMTEST_CHECK_UNQUOTED([
 RPMDB_INIT
 
@@ -772,6 +782,7 @@ RPMTEST_CLEANUP
 # Test pre-built corrupted package verification (corrupted header)
 AT_SETUP([rpmkeys -Kv <corrupted signed> 2.1])
 AT_KEYWORDS([rpmkeys digest signature])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMTEST_CHECK([
 RPMDB_INIT
 
@@ -806,6 +817,7 @@ RPMTEST_CLEANUP
 # Test pre-built corrupted package verification (corrupted header)
 AT_SETUP([rpmkeys -Kv <corrupted signed> 2.2])
 AT_KEYWORDS([rpmkeys digest signature])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMTEST_CHECK([
 RPMDB_INIT
 
@@ -841,6 +853,7 @@ RPMTEST_CLEANUP
 # Test pre-built corrupted package verification (corrupted payload)
 AT_SETUP([rpmkeys -Kv <corrupted signed> 3])
 AT_KEYWORDS([rpmkeys digest signature])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMTEST_CHECK([
 RPMDB_INIT
 
@@ -879,6 +892,7 @@ RPMTEST_CLEANUP
 # Test pre-built corrupted package verification (corrupted payload)
 AT_SETUP([rpmkeys -Kv <corrupted signed> 4])
 AT_KEYWORDS([rpmkeys digest signature])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMTEST_CHECK([
 RPMDB_INIT[(
 
@@ -918,6 +932,7 @@ RPMTEST_CLEANUP
 # Test --addsign
 AT_SETUP([rpmsign --addsign])
 AT_KEYWORDS([rpmsign signature])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMDB_INIT
 gpg2 --import ${RPMTEST}/data/keys/*.secret
 # Our keys have no passphrases to be asked, silence GPG_TTY warning
@@ -1041,6 +1056,7 @@ RPMTEST_CLEANUP
 # Test --delsign
 AT_SETUP([rpmsign --delsign])
 AT_KEYWORDS([rpmsign signature])
+AT_SKIP_IF([test x$PGP = xdummy])
 RPMTEST_CHECK([
 RPMDB_INIT
 
@@ -1060,4 +1076,25 @@ POST-DELSIGN
 /tmp/hello-2.0-1.x86_64-signed.rpm:
 ],
 [])
+RPMTEST_CLEANUP
+
+AT_SETUP([without openpgp])
+AT_KEYWORDS([dummy])
+AT_SKIP_IF([test x$PGP != xdummy])
+RPMTEST_CHECK([
+runroot rpmkeys --import /data/keys/rpm.org-rsa-2048-test.pub
+],
+[1],
+[],
+[error: /data/keys/rpm.org-rsa-2048-test.pub: key 1 not an armored public key.
+])
+
+RPMTEST_CHECK([
+runroot rpm -qp /data/RPMS/hello-2.0-1.x86_64-signed.rpm
+],
+[1],
+[],
+[error: /data/RPMS/hello-2.0-1.x86_64-signed.rpm: Header RSA signature: BAD (package tag 268: invalid OpenPGP signature: RPM was compiled without OpenPGP support)
+error: /data/RPMS/hello-2.0-1.x86_64-signed.rpm: not an rpm package (or package manifest)
+])
 RPMTEST_CLEANUP
